+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Extensions
 - `CharacterSet` `unified` property is added
 - `Calendar` can now calculate number of days in month from year and month as arguments
 - `SizeConfiguration` is renamed to `MinIdealMaxSizes`
+- App info properties are moved from `UIApplication` to `Bundle`, and support is added for all platforms
 
 #### [3.3.1(24)](https://github.com/VakhoKontridze/VCore/releases/tag/3.3.1) — *2022 06 08*